削除するときに確認ダイアログ表示 #1

+++ app/routes/tasks.js
@@ -7,7 +7,11 @@ export default Ember.Route.extend({
 
   actions: {
     deleteTask(task) {
-      task.destroyRecord();
+      let confirmation = confirm('削除しますか？');
+
+      if (confirmation) {
+        task.destroyRecord();
+      }
     }
   }
 });
